Move some functions to tools file
