+++ python/BERT_Analyze.py
@@ -2,59 +2,12 @@
 
 import os
 import glob
-import re
 from BERT_Plot import plot, plotMultiple
-
-# return list of TAP0 settings that had errors
-def findErrors(input_file):
-    f = open(input_file, 'r')
-    # list of TAP0 settings that had errors
-    errors = []
-    for line in f:
-        # check for errors
-        if "returncode=1" in line:
-            # get all numbers in string
-            numbers = re.findall(r'\d+', line)
-            # first number is TAP0 setting
-            z = int(numbers[0])
-            #print("ERROR: {0}".format(line), end='')
-            errors.append(z)
-    f.close()
-    return errors
-
-def getData(input_file):
-    # check for errors
-    printError = False
-    errors = findErrors(input_file)
-    if errors and printError:
-        print("ERROR for {0}".format(input_file))
-        print("There were errors for these TAP0 settings: {0}".format(errors))
-        print("These will be skipped.")
-    f = open(input_file, 'r')
-    x_values = []
-    y_values = []
-    for line in f:
-        # TAP0 DAC Setting (x values)
-        if "CML_TAP0_BIAS" in line:
-            array = line.split()
-            # must remove " before using int()
-            x = int(array[-1].replace('"', ''))
-            # skip the x value if there were errors
-            if x not in errors:
-                x_values.append(x)
-        # Total error counter (y values)
-        if "Final counter" in line:
-            # get all numbers in string
-            numbers = [int(s) for s in line.split() if s.isdigit()]
-            y = numbers[0]
-            y_values.append(y)
-    
-    f.close()
-    return [x_values, y_values]
+from tools import getBERTData
 
 def analyze(input_file, plot_dir, output_file):
     debug = False
-    data = getData(input_file)
+    data = getBERTData(input_file)
     x_values = data[0]
     y_values = data[1]
     # check for the same number of x and y values
@@ -110,7 +63,7 @@ def analyzeScans():
 # add entry to list of inputs
 # each entry is a dictionary
 def addEntry(input_list, input_file, label):
-    data = getData(input_file)
+    data = getBERTData(input_file)
     entry = {}
     entry["x_values"] = data[0]
     entry["y_values"] = data[1]
